+++ repo/src/main/java/com/parashift/onlyoffice/CallBack.java
@@ -2,6 +2,8 @@ package com.parashift.onlyoffice;
 
 import org.alfresco.model.ContentModel;
 import org.alfresco.repo.policy.BehaviourFilter;
+import org.alfresco.repo.security.authentication.AuthenticationUtil;
+import org.alfresco.repo.transaction.RetryingTransactionHelper.RetryingTransactionCallback;
 import org.alfresco.service.cmr.lock.LockService;
 import org.alfresco.service.cmr.lock.LockStatus;
 import org.alfresco.service.cmr.lock.LockType;
@@ -10,15 +12,15 @@ import org.alfresco.service.cmr.repository.ContentService;
 import org.alfresco.service.cmr.repository.MimetypeService;
 import org.alfresco.service.cmr.repository.NodeRef;
 import org.alfresco.service.cmr.repository.NodeService;
+import org.alfresco.service.transaction.TransactionService;
 import org.apache.commons.lang.exception.ExceptionUtils;
-import org.json.JSONException;
+import org.json.JSONArray;
 import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.extensions.webscripts.AbstractWebScript;
-import org.springframework.extensions.webscripts.WebScriptException;
 import org.springframework.extensions.webscripts.WebScriptRequest;
 import org.springframework.extensions.webscripts.WebScriptResponse;
 import org.springframework.stereotype.Component;
@@ -72,10 +74,16 @@ public class CallBack extends AbstractWebScript {
     @Autowired
     Converter converterService;
 
+    @Autowired
+    TransactionService transactionService;
+
     private Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Override
     public void execute(WebScriptRequest request, WebScriptResponse response) throws IOException {
+
+        Integer code = 0;
+        Exception error = null;
 
         logger.debug("Received JSON Callback");
         try {
@@ -94,15 +102,11 @@ public class CallBack extends AbstractWebScript {
                 }
 
                 if (token == null || token == "") {
-                    response.setStatus(403);
-                    response.getWriter().write("{\"error\": 0, \"message\": \"Expected JWT\"}");
-                    return;
+                    throw new SecurityException("Expected JWT");
                 }
 
                 if (!jwtManager.verify(token)) {
-                    response.setStatus(403);
-                    response.getWriter().write("{\"error\": 0, \"message\": \"Wrong JWT\"}");
-                    return;
+                    throw new SecurityException("JWT verification failed");
                 }
 
                 JSONObject bodyFromToken = new JSONObject(new String(Base64.getUrlDecoder().decode(token.split("\\.")[1]), "UTF-8"));
@@ -116,14 +120,76 @@ public class CallBack extends AbstractWebScript {
 
             String[] keyParts = callBackJSon.getString("key").split("_");
             NodeRef nodeRef = new NodeRef("workspace://SpacesStore/" + keyParts[0]);
+            String hash = (String) nodeService.getProperty(nodeRef, Util.EditingHashAspect);
+            String queryHash = request.getParameter("cb_key");
 
+            if (hash == null || queryHash == null || !hash.equals(queryHash)) {
+                throw new SecurityException("Security hash verification failed");
+            }
+
+            String username = null;
+
+            if (callBackJSon.has("users")) {
+                JSONArray array = callBackJSon.getJSONArray("users");
+                if (array.length() > 0) {
+                    username = (String) array.get(0);
+                }
+            }
+
+            if (username == null && callBackJSon.has("actions")) {
+                JSONArray array = callBackJSon.getJSONArray("actions");
+                if (array.length() > 0) {
+                    username = ((JSONObject) array.get(0)).getString("userid");
+                }
+            }
+
+            if (username != null) {
+                AuthenticationUtil.setRunAsUser(username);
+            } else {
+                throw new SecurityException("No user information");
+            }
+            Boolean reqNew = transactionService.isReadOnly();
+            transactionService.getRetryingTransactionHelper()
+                .doInTransaction(new ProccessRequestCallback(callBackJSon, nodeRef), reqNew, reqNew);
+            AuthenticationUtil.clearCurrentSecurityContext();
+
+        } catch (SecurityException ex) {
+            code = 403;
+            error = ex;
+        } catch (Exception ex) {
+            code = 500;
+            error = ex;
+        }
+
+        if (error != null) {
+            response.setStatus(code);
+            logger.error(ExceptionUtils.getFullStackTrace(error));
+
+            response.getWriter().write("{\"error\":1, \"message\":\"" + error.getMessage() + "\"}");
+        } else {
+            response.getWriter().write("{\"error\":0}");
+        }
+    }
+
+    private class ProccessRequestCallback implements RetryingTransactionCallback<Object> {
+
+        private JSONObject callBackJSon;
+        private NodeRef nodeRef;
+
+        public ProccessRequestCallback(JSONObject json, NodeRef node) {
+            callBackJSon = json;
+            nodeRef = node;
+        }
+
+        @Override
+        public Object execute() throws Throwable {
             //Status codes from here: https://api.onlyoffice.com/editors/editor
-
-            int saved = 0;
             switch(callBackJSon.getInt("status")) {
                 case 0:
                     logger.error("ONLYOFFICE has reported that no doc with the specified key can be found");
                     lockService.unlock(nodeRef);
+                    logger.info("removing prop");
+                    nodeService.removeProperty(nodeRef, Util.EditingHashAspect);
                     break;
                 case 1:
                     if(lockService.getLockStatus(nodeRef).equals(LockStatus.NO_LOCK)) {
@@ -137,51 +203,49 @@ public class CallBack extends AbstractWebScript {
                 case 2:
                     logger.debug("Document Updated, changing content");
                     lockService.unlock(nodeRef);
-                    if (!updateNode(nodeRef, callBackJSon.getString("url")))
-                    {
-                        saved = 1;
-                    }
+                    logger.info("removing prop");
+                    nodeService.removeProperty(nodeRef, Util.EditingHashAspect);
+                    updateNode(nodeRef, callBackJSon.getString("url"));
                     break;
                 case 3:
                     logger.error("ONLYOFFICE has reported that saving the document has failed");
                     lockService.unlock(nodeRef);
+                    logger.info("removing prop");
+                    nodeService.removeProperty(nodeRef, Util.EditingHashAspect);
                     break;
                 case 4:
                     logger.debug("No document updates, unlocking node");
                     lockService.unlock(nodeRef);
+                    logger.info("removing prop");
+                    nodeService.removeProperty(nodeRef, Util.EditingHashAspect);
                     break;
             }
-
-            response.getWriter().write("{\"error\":" + saved + "}");
-        } catch (JSONException ex) {
-            throw new WebScriptException("Unable to deserialize JSON: " + ex.getMessage());
+            return null;
         }
     }
 
-    private boolean updateNode(NodeRef nodeRef, String url) {
+    private void updateNode(NodeRef nodeRef, String url) throws Exception {
         logger.debug("Retrieving URL:{}", url);
-            ContentData contentData = (ContentData) nodeService.getProperty(nodeRef, ContentModel.PROP_CONTENT);
-            String mimeType = contentData.getMimetype();
+        ContentData contentData = (ContentData) nodeService.getProperty(nodeRef, ContentModel.PROP_CONTENT);
+        String mimeType = contentData.getMimetype();
 
-            if (converterService.shouldConvertBack(mimeType)) {
-                try {
-                    logger.debug("Should convert back");
-                    url = converterService.convert(nodeRef.getId(), "docx", mimetypeService.getExtension(mimeType), url);
-                } catch (Exception e) {
-                    logger.error(ExceptionUtils.getFullStackTrace(e));
-                    return false;
-                }
-            }
-
+        if (converterService.shouldConvertBack(mimeType)) {
             try {
-                checkCert();
-                InputStream in = new URL( url ).openStream();
-                contentService.getWriter(nodeRef, ContentModel.PROP_CONTENT, true).putContent(in);
-            } catch (IOException e) {
-                logger.error(ExceptionUtils.getFullStackTrace(e));
-                return false;
+                logger.debug("Should convert back");
+                url = converterService.convert(nodeRef.getId(), "docx", mimetypeService.getExtension(mimeType), url);
+            } catch (Exception e) {
+                throw new Exception("Error while converting document back to original format: " + e.getMessage(), e);
             }
-        return true;
+        }
+
+        try {
+            checkCert();
+            InputStream in = new URL( url ).openStream();
+            contentService.getWriter(nodeRef, ContentModel.PROP_CONTENT, true).putContent(in);
+        } catch (IOException e) {
+            logger.error(ExceptionUtils.getFullStackTrace(e));
+            throw new Exception("Error while downloading new document version: " + e.getMessage(), e);
+        }
     }
 
     private void checkCert() {